+++ lua/raccoon/config.lua
@@ -27,7 +27,7 @@ M.defaults = {
   clone_root = vim.fs.joinpath(vim.fn.stdpath("data"), "raccoon", "repos"),
   pull_changes_interval = 300,
   commit_viewer = {
-    grid = { rows = 2, cols = 2 },
+    grid = { rows = 2, cols = 1 },
     base_commits_count = 20,
     sidebar_width = 50,
   },
@@ -190,7 +190,7 @@ function M.create_default()
     clone_root = vim.fs.joinpath(vim.fn.stdpath("data"), "raccoon", "repos"),
     pull_changes_interval = 300,
     commit_viewer = {
-      grid = { rows = 2, cols = 2 },
+      grid = { rows = 2, cols = 1 },
       base_commits_count = 20,
     },
   }